pyproject.toml
[tool.poetry]
name = "micro-registry"
version = "0.2.6"
description = "A Python library for managing and loading class instances from modules and YAML configurations."
authors = ["Aleksander Stanik <aleksander.stanik@hammerheadesengineers.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0.2"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
python-semantic-release = "^9.14.0"
black = "^24.10.0"
isort = "^5.13.2"
commitizen = "^3.31.0"
flake8 = "^7.1.1"

[tool.poetry.scripts]
micro-app = "micro_app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variables = [
    "micro_registry/__init__.py:__version__"
]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]

commit_message = "Release v{version}"
build_command = "pip install poetry && poetry build"

[tool.semantic_release.branches.main]
match = "main"
prerelease = false
prerelease_token = "rc"

[tool.semantic_release.branches.alpha]
match = "^(feat|fix|perf)/.+"
prerelease = true
prerelease_token = "alpha"

[tool.semantic_release.branches.develop]
match = "develop"
prerelease = true
prerelease_token = "rc"

[tool.semantic_release.remote]
type = "github"
token = { env = "GH_TOKEN" }

[tool.semantic_release.publish]
upload_to_vcs_release = true